requirements.txt
tensorflow-intel==2.16.1
pygame==2.5.2
gym==0.26.2